@@ -1,13 +1,8 @@
 package com.mariesto.paymentservice.persistence.entity;
 
 import java.sql.Timestamp;
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
+
+import jakarta.persistence.*;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 import com.mariesto.paymentservice.constant.PaymentStatus;
@@ -16,12 +11,12 @@ import lombok.Getter;
 import lombok.Setter;
 
 @Entity
-@Table (name = "payments")
+@Table(name = "payments")
 @Getter
 @Setter
 public class Payment {
     @Id
-    @GeneratedValue (strategy = GenerationType.IDENTITY)
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     private String paymentReferenceId;
