@@ -1,14 +1,12 @@
 package com.mariesto.userservice.persistence.entity;
 
 import java.sql.Timestamp;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
+
+import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
+import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.UpdateTimestamp;
 
 @Entity
 @Table(name = "users")
@@ -17,21 +15,18 @@ import lombok.Setter;
 public class User {
 
     @Id
-    @GeneratedValue (strategy = GenerationType.IDENTITY)
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column
     private String userId;
 
-    @Column
     private String username;
 
-    @Column
     private String phoneNumber;
 
-    @Column
+    @CreationTimestamp
     private Timestamp createdAt;
 
-    @Column
+    @UpdateTimestamp
     private Timestamp updatedAt;
 }
